Add support for multiple endpoints
This commits adds a new `endpoints` parameter for the `BaseClient` class.
This parameter, which must be used alternatively to `host` and `port`,
can be set to a list of EtcdEndpoint, which is a simple data class with
`address` and `port` attributes.

A new decorator is  present: `retry all hosts`. This decorator if
applied to a function of a class descendant from `BaseClient` has the
function calls retried against all the `endpoints` if an exception
occours.
If one of the tries returns with no errors the return value is propagated;
if all the tries throws exceptions of the same type the first one is
propagated; if they throw different exceptions an Etcd3Exception is
thrown. All the failed tries are logged.