Rework tests to run against an etcd cluster
A new utility class `EtcdCluster` is added to manage the lifecycle of
a etcd cluster run in containers.
This class replaces the methods in `docker_cli` and `etcd_go_cli`.

A set of fixtures(`etcd_cluster`, `etcd_cluster_ssl`, `client` and
`io_client` manages the lifecycle of the test resources.

All the tests are set to run using the `BaseClient`'s `endpoint` parameter,
additional tests have been added to test `host` and `port`.