rework failover logic
every call to a decorated function will now try the call againts the
current in-use endpoint before trying to failover.

failover will be tried only if the exception thrown is in
failover_exceptions and the api method called is in failover_whitelist.

the failover process is blocking: Client and AioClient defines their own
Lock compatible object